rename and only run setup when imported into test
