+++ client/src/test/setup.ts
@@ -1,2 +1,1 @@
 import '@testing-library/jest-dom/vitest';
-import './msw-setup';
